Bug fixed in the TestBoxTest class and dependecies optimized
